feat: useCourseDetail 커스텀 훅 추가
